feat: [M467] implement bleaching summary stats on collect record page including refactoring submitted page to reuse some of the code. Implemented forgotten percent cover observations table for submitted page.
