+++ src/components/BleachingColoniesBleachedSummaryStats/BleachingColoniesBleachedSummaryStats.js
@@ -1,0 +1,89 @@
+import React from 'react'
+import { observationsColoniesBleachedPropType } from '../../App/mermaidData/mermaidDataProptypes'
+import { Td, Th, Tr } from '../generic/Table/table'
+import { ObservationsSummaryStats } from '../pages/collectRecordFormPages/CollectingFormPage.Styles'
+
+const BleachincColoniesBleachedSummaryStats = ({ observationsColoniesBleached }) => {
+  const getTotalOfColonies = () => {
+    const totals = observationsColoniesBleached.map((item) => {
+      return (
+        Number(item.count_20) +
+        Number(item.count_50) +
+        Number(item.count_80) +
+        Number(item.count_100) +
+        Number(item.count_dead) +
+        Number(item.count_normal) +
+        Number(item.count_pale)
+      )
+    })
+
+    return totals.reduce((acc, currentVal) => acc + currentVal, 0)
+  }
+
+  const getTotalOfCoralGenera = () => {
+    const attributeIds = observationsColoniesBleached.map((item) => item.attribute)
+    const uniqueAttributeIds = [...new Set(attributeIds)]
+
+    return uniqueAttributeIds.length
+  }
+
+  const getPercentageOfColonies = (colonyType) => {
+    let totals = 0
+
+    if (colonyType === 'bleached') {
+      totals = observationsColoniesBleached.map((item) => {
+        return (
+          Number(item.count_20) +
+          Number(item.count_50) +
+          Number(item.count_80) +
+          Number(item.count_100) +
+          Number(item.count_dead)
+        )
+      })
+    } else {
+      totals = observationsColoniesBleached.map((item) => item[colonyType])
+    }
+
+    return (
+      (totals.reduce((acc, currentVal) => acc + currentVal, 0) / getTotalOfColonies()) *
+      100
+    ).toFixed(1)
+  }
+
+  return (
+    <ObservationsSummaryStats>
+      <tbody>
+        <Tr>
+          <Th>Total number of colonies</Th>
+          <Td>{getTotalOfColonies()}</Td>
+        </Tr>
+        <Tr>
+          <Th>Total number of coral genera</Th>
+          <Td>{getTotalOfCoralGenera()}</Td>
+        </Tr>
+        <Tr>
+          <Th>% Normal colonies</Th>
+          <Td>{getPercentageOfColonies('count_normal')}</Td>
+        </Tr>
+        <Tr>
+          <Th>% Pale colonies</Th>
+          <Td>{getPercentageOfColonies('count_pale')}</Td>
+        </Tr>
+        <Tr>
+          <Th>% Bleached colonies</Th>
+          <Td>{getPercentageOfColonies('bleached')}</Td>
+        </Tr>
+      </tbody>
+    </ObservationsSummaryStats>
+  )
+}
+
+BleachincColoniesBleachedSummaryStats.propTypes = {
+  observationsColoniesBleached: observationsColoniesBleachedPropType,
+}
+
+BleachincColoniesBleachedSummaryStats.defaultProps = {
+  observationsColoniesBleached: [],
+}
+
+export default BleachincColoniesBleachedSummaryStats
