--- src/components/BleachingObservationSummaryStats/index.js
@@ -1,3 +1,0 @@
-import BleachingObservationSummaryStats from './BleachingObservationSummaryStats'
-
-export default BleachingObservationSummaryStats
