+++ src/components/pages/submittedRecordPages/SubmittedBleaching/BleachingColoniesBleachedObservations/index.js
@@ -1,0 +1,3 @@
+import BleachingColoniesBleachedObservations from './BleachingColoniesBleachedObservations'
+
+export default BleachingColoniesBleachedObservations
